requirements.txt
python-telegram-bot==11.1.0
timeout-decorator==0.4.0
